🖥️ ui: extract toast presentation component
